working put car method

+++ server/server/src/main/java/ghart/space/server/car/CarNotFoundException.java
@@ -2,7 +2,7 @@ package ghart.space.server.car;
 
 class CarNotFoundException extends RuntimeException {
 
-    CarNotFoundException(Long id) {
+    CarNotFoundException(Integer id) {
     super("Could not find car " + id);
     }
 }
